tests: Create dedicated secondary command buffer tests
